Fix salt alignment in a bunch of formats, see #3932

+++ src/andotp_fmt_plug.c
@@ -41,7 +41,7 @@ john_register_one(&fmt_andotp);
 #define BINARY_SIZE             0
 #define BINARY_ALIGN            1
 #define SALT_SIZE               sizeof(struct custom_salt)
-#define SALT_ALIGN              sizeof(uint64_t)
+#define SALT_ALIGN              sizeof(uint32_t)
 #define MIN_KEYS_PER_CRYPT      1
 #define MAX_KEYS_PER_CRYPT      4  // tuned on i7-7820HQ
